Some clean-up again

+++ tools/src/main/java/org/esa/cci/sst/common/calculator/Accumulator.java
@@ -26,5 +26,5 @@ package org.esa.cci.sst.common.calculator;
  */
 public interface Accumulator {
 
-    long getSampleCount();
+    int getSampleCount();
 }
